third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/scroll-to-fragid/scroll-position-vertical-lr-expected.txt

This is a testharness.js-based test.
FAIL Fragment Navigation: Scroll to block start position in vertical-lr writing mode assert_equals: Scroll to the left border edge of #test expected 14 but got 0
Harness: the test ran to completion.
